kernel: make FreeRTOS private to kernel/freertos
third_party/freertos now exports its include directories only to the
shim, so any other include of a FreeRTOS header fails to compile.
FreeRTOSConfig.h drops the firmware headers it used to pull in for the
whole tree and the extra TLS slots the syscall code no longer needs;
the call sites that relied on the transitive includes name their
headers. The tests and host tools drop the FreeRTOS include paths and
the last stale fakes. The SiFli IPC and nrfx glue headers, which
declared vPortEnterCritical() themselves, use pbl_irq_lock().

The moddable submodule's Pebble platform headers still call FreeRTOS
directly, so the libxs library and the one applib file that includes
them keep a private view of the headers until the submodule is ported.

Verified on qemu_gabbro (boots to the watchface, no asserts), asterix
and obelix builds, and the unit test suite.
