+++ lib/os/README.md
@@ -1,11 +1,11 @@
 # libOS
 
-libOS is a helper library that makes developing software on top of FreeRTOS on ARM easier.
-It is used by and built for the main FW and the Dialog Bluetooth FW.
+libOS holds the platform helpers the firmware and the Dialog Bluetooth FW share:
+assertions, the allocator hooks and the Cortex-M MCU state helpers. Threading
+lives in `kernel/`.
 
 ## Dependencies:
 
 - libc
 - libutil
-- FreeRTOS
 - A handful of platform specific functions, see platform.c
