+++ tests/fakes/fake_events.c
@@ -4,7 +4,6 @@
 #include "fake_events.h"
 #include "kernel/pbl_malloc.h"
 
-#include "projdefs.h"
 
 static PebbleEvent s_last_pebble_event;
 static uint32_t s_fake_event_count = 0;
