+++ tests/fw/services/test_light.c
@@ -27,9 +27,6 @@
 #include "stubs_mutex.h"
 #include "stubs_rtc.h"
 
-void vTaskDelay(uint32_t ticks) {
-}
-
 // the time that the backlight remains on but there is zero user interaction
 extern const uint32_t INACTIVE_LIGHT_TIMEOUT_MS;
 // the time duration of the fade out
